better structure for the program
