+++ osa2/puhelinluettelo/db.json
@@ -36,24 +36,14 @@
       "number": "22"
     },
     {
-      "id": "7744",
-      "name": "bee",
-      "number": "123"
+      "id": "cf0e",
+      "name": "uus",
+      "number": "12345"
     },
     {
-      "id": "5201",
-      "name": "cee",
-      "number": "321"
-    },
-    {
-      "id": "0601",
-      "name": "fee",
-      "number": "e2"
-    },
-    {
-      "id": "360b",
-      "name": "2355636",
-      "number": "aegaegt"
+      "id": "bf83",
+      "name": "ii",
+      "number": "1"
     }
   ]
 }
