+++ osa2/puhelinluettelo/src/components/Filter.jsx
@@ -1,0 +1,8 @@
+const Filter = ({findPerson, handleSearch}) => {
+    return(
+    <div>filter shown with <input value={findPerson} onChange={handleSearch}></input></div>
+    )
+  }
+
+  
+export default Filter
