+++ osa2/puhelinluettelo/src/components/Personform.jsx
@@ -1,0 +1,15 @@
+const PersonForm = ({newName, newPhone, onNameChange, onPhoneChange, onSubmit}) => {
+    return( 
+     <form onSubmit={onSubmit}>
+           <div>
+             name: <input value={newName} onChange={onNameChange}/>
+           </div>
+           <div>number: <input value={newPhone} onChange={onPhoneChange}/></div>
+           <div>
+             <button type="submit">add</button>
+           </div>
+         </form>
+    )
+   }
+
+   export default PersonForm
