+++ osa2/puhelinluettelo/src/components/Persons.jsx
@@ -1,0 +1,11 @@
+const Persons = ({persons}) => {
+    return(
+    <ul>
+        {persons.map(person => 
+            <li key={person.name}>{person.name} {person.number}</li>
+          )}
+        </ul>
+    )
+  }
+
+  export default Persons
